Delete Actor & Director updated

+++ WebApi/Application/ActorOperations/Commands/DeleteActor/DeleteActorCommand.cs
@@ -18,6 +18,10 @@ public class DeleteActorCommand
         if(actor is null)
             throw new InvalidOperationException("The actor doesn't exist.");
 
+        var movies = _dbContext.MovieActors.Where(x => x.ActorId == ActorId).Any();
+        if(movies)
+            throw new InvalidOperationException("The actor has movie/s. To delete the actor you must first delete the actor's movie/s.");
+
         _dbContext.Actors.Remove(actor);
         _dbContext.SaveChanges();
     }
