+++ WebApi/Application/DirectorOperations/Commands/DeleteDirector/DeleteDirectorCommand.cs
@@ -18,6 +18,10 @@ public class DeleteDirectorCommand
         if(director is null)
             throw new InvalidOperationException("The director doesn't exist.");
 
+        var movies = _dbContext.Movies.Where(x => x.DirectorId == directorId).Any();
+        if(movies)
+            throw new InvalidOperationException("The director has movie/s. To delete the director you must first delete the director's movie/s.");
+
         _dbContext.Directors.Remove(director);
         _dbContext.SaveChanges();    
     }
